Fixed a subtle and invisible bug with URLs.  We were unescaping punctuation, and then fetching pages from the server using the unescaped URLs.  This led to problems later, when I wanted to unescape spaces, which aren't URL-legal.
Unescaped spaces, so the graph is prettier.

Added quotes around dot output.

Added URLs on the nodes, so we can generate imagemaps.

Changed the output file from a named file to stdout, to be Unixier.

Made some list code a little cleaner by using list comprehension syntax.

Added a shell script wrapper to generate a PNG and an imagemap.


git-svn-id: https://saturnvalley.org/svn/lexgraph/trunk@466 e8952cfc-6410-0410-bd16-81fef15e32ef